fix(opensearchservice): access denied when creating a new domain in regions without cognito support (#21395)
Closes #21192

Currently the `CognitoOptions` attribute is always populated in the CloudFormation output when creating a new OpenSearch Domain, even if `cognitoDashboardsAuth` is not passed into the constructor.

Including this attribute, even with just the default body of `{ "Enabled": false }`, causes deployments to fail in regions where Cognito isn't supported (e.g. `cn-north-1`), with the message: `You don't have permissions to integrate with Cognito. Contact your admin if you need help`. This change fixes this behavior by completely omitting the attribute if `cognitoDashboardsAuth` is not passed into the constructor.

I also updated the package.json file to fix several build warnings, and the README and default Rosetta fixture to fix a build error and make the docs align with the style recommendations in the contribution guidelines.

**NOTE:** Current snapshots are updated after running the integration tests, since the existing templates should no longer have `CognitoOptions` populated. However, the snapshot for the newly added integration test has this attribute populated.

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [x] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [x] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*