fix(aws-ecs): missing iam policies for firelens cloudwatch_logs log driver
We recently migrated our FireLens log driver from using `cloudwatch` to using the recommended `cloudwatch_logs`.

`cloudwatch_logs` is the newer and more high performance version as detailed in this guidance from AWS https://github.com/aws/amazon-cloudwatch-logs-for-fluent-bit#new-higher-performance-core-fluent-bit-plugin

However, when doing that, we lost the automatic configuration of the IAM policies to write to CloudWatch which broke our logging.

Looks like the CDK only auto adds CloudWatch IAM policies only for the `cloudwatch` plugin opposed to both the older `cloudwatch` plugin and newer and recommended `cloudwatch_logs` plugin.

Policies should be added to both so that customer's logging doesn't break when migrating to the newer plugin.