feat(ecs-patterns) expose multiple target groups, listeners, and load balancers  (#20770)
This commit will expose targetGroups, listeners, and loadBalancers properties allowing developers access to those resources.

The downside to this is that now we have targetGroups, targetGroup, listeners, listener, and loadBalancers, loadBalancer exposes as a property. In my opinion this can cause confusion for developers.
 
** Potential breaking change **
When trying to protect targetGroup, listener, and loadBalancer I discovered that this would cause breaking changes on down developers who use them to reference `node.Id` and other attributes. 

Only thing I could do was mark them as `deprecated` as you get the same functionality on the new properties. Open to feedback on this approach and will update the README based on the direction we go.

I need to update the network Integ tests some more, but looking for feedback on the above

closes #14735
----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [x] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [x] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*