docs(ssm): document how to use valueFromLookup for specific value formats
`StringParameter.valueFromLookup` can be used to retrieve a string
value, but that value could be in any format and there are times where
it is provided as input to properties that require a specific format
(i.e. arn format). Because of the way that lookups are resolved, it is
possible for the initial value to be the dummy value
`dummy-value-for-${parameterName}` which might cause synth errors.

Since there is no way for the CDK to know _how_ you will use the
returned value, we can't really add logic to specifically handle edge
cases. For example, we could have `valueFromLookup` always return a
token, but then it would no longer be able to be used in cases where a
`string` is required. See https://github.com/aws/aws-cdk/issues/8699#issuecomment-976159825
for a good analysis.

This PR adds documentation instructing users how to handle these use
cases.

closes #8699