fix(route53-targets): InterfaceVpcEndpoint target requires a InterfaceVpcEndpoint
The `InterfaceVpcEndpointTarget` currently accepts an imported endpoint
`IInterfaceVpcEndpoint`, but the actual implementation requires access
to the `CfnVPCEndpoint` resource which is only available from the
`InterfaceVpcEndpoint`.

This appears as breaking, but I don't think it should be considered as
breaking since you can currently _only_ pass a `InterfaceVpcEndpoint`.

I thought about adding support for `IInterfaceVpcEndpoint`, but that
would require updating the `fromxxx` method to take the route53 zone and
the DNS name for the endpoint. If you need to know those values to use
this construct then you might as well just create a normal Route53
record.

closes #10432