fix(route53-targets): InterfaceVpcEndpointTarget incorrectly accepts an imported endpoint (#21523)
The `InterfaceVpcEndpointTarget` currently accepts an imported endpoint
`IInterfaceVpcEndpoint`, but the actual implementation requires access
to the `CfnVPCEndpoint` resource which is only available from the
`InterfaceVpcEndpoint`.

This appears as breaking, but I don't think it should be considered as
breaking since you can currently _only_ pass a `InterfaceVpcEndpoint`.

I thought about adding support for `IInterfaceVpcEndpoint`, but that
would require updating the `fromxxx` method to take the route53 zone and
the DNS name for the endpoint. If you need to know those values to use
this construct then you might as well just create a normal Route53
record.

closes #10432


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*