fix(ec2): unavailable kernel version for Amazon Linux 2022 fails deployment
Amazon Linux 2022 is in preview release

It appears that previous kernel versions are not retained in this phase.
While existing deployments will still work, it is not possible to create/update
instances with an old kernel version.

This change will change the kernel version for all customers to the default
version, unless they have opted into a specific version. Since the AL22 is in
preview, we deem this an acceptable breaking change.

https://docs.aws.amazon.com/linux/al2022/ug/naming-and-versioning.html

Fixes #21011