chore(cloud9): fix integ test that cannot pass (#21561)
This integ test cannot pass using `CONNECT_SSM`  because it requires setup of resources
ahead of time that the L1 construct does not handle. Because roles use eventual consistency, 
just adding the role to the test isn't a good solution. `CONNECT_SSH` works just fine.

A note has been added to the `ConnectionType.CONNECT_SSM` enum that a role and instance
profile must be setup beforehand if it is selected. Eventually if/when an L2 construct is created 
we should handle this setup.


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*