fix(codepipeline-actions): ecr source action doesn't trigger the pipeline
When the ECR source action was initially released, ECR did not yet
integrate with EventBridge so it was necessary to use CloudTrail.
Since then ECR was updated to integrate with EventBridge so relying on
CloudTrail is no longer necessary.

This PR changes the event rule that triggers the pipeline to use the
`ECR Image Action` instead of the CloudTrail event.

This change does lead to the `AWS::Events::Rule` resource being deleted
and recreated, but that should not cause any issues since there is no
state involved. I tested this by using the integ test and manually
validated that pushing an image to the ecr repo triggers the pipeline.
The steps are outlined in the `integ.pipeline-ecr-source.ts` file.

fix #10901