fix(codepipeline-actions): ecr source action doesn't trigger the pipeline (#21580)
When the ECR source action was initially released, ECR did not yet
integrate with EventBridge so it was necessary to use CloudTrail.
Since then ECR was updated to integrate with EventBridge so relying on
CloudTrail is no longer necessary.

This PR changes the event rule that triggers the pipeline to use the
`ECR Image Action` instead of the CloudTrail event.

This change does lead to the `AWS::Events::Rule` resource being deleted
and recreated, but that should not cause any issues since there is no
state involved. I tested this by using the integ test and manually
validated that pushing an image to the ecr repo triggers the pipeline.
The steps are outlined in the `integ.pipeline-ecr-source.ts` file.

fix #10901


----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [x] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [x] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*