fix(route53): misleading error message in `fromLookup` if `domainName` is undefined (#21596)
If TypeScript compiler option `strictPropertyInitialization` is set to `false`, it is possible to pass an undefined value to the `fromLookup` method in `HostedZone`. Currently the error message `Cannot read property 'endsWith' of undefined` will be displayed. This error message does not indicate the root cause, so it is difficult to fix the problem.

This fix introduces an additional check that shows an appropriate error message.

Fixes #10053.

----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*