fix(integ-runner): ignoring asset changes doesn't work with new style assets
The integ-runner has the ability to ignore changes to assets since asset
changes can lead to a lot of snapshot failures and we don't always need
to re-deploy the integ test when this occurs. The `canonicalizeTemplate`
function only worked with the v1
(`@aws-cdk/core:newStyleStackSynthesis:false`) assets which were stored
in CFN parameters.

This PR updates the logic to also account for the new style assets.
Because we now have a list of the actual assets in the asset manifest
the logic simply reads the list of assets from the manifest and replaces
substitutes those values.

I also refactored some things:
- The `canonicalizeTemplate` function was moved into a private method of
 `IntegSnapshotRunner` since that is the only place it is used.
- The CloudAssemblyManifest asset methods were refactored so that we
  could either return the asset location or the asset id