fix(integ-runner): ignoring asset changes doesn't work with new style assets (#21638)
The integ-runner has the ability to ignore changes to assets since asset
changes can lead to a lot of snapshot failures and we don't always need
to re-deploy the integ test when this occurs. The `canonicalizeTemplate`
function only worked with the v1
(`@aws-cdk/core:newStyleStackSynthesis:false`) assets which were stored
in CFN parameters.

This PR updates the logic to also account for the new style assets.
Because we now have a list of the actual assets in the asset manifest
the logic simply reads the list of assets from the manifest and replaces
substitutes those values.

I also refactored some things:
- The `canonicalizeTemplate` function was moved into a private method of
 `IntegSnapshotRunner` since that is the only place it is used.
- The CloudAssemblyManifest asset methods were refactored so that we
  could either return the asset location or the asset id


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*