feat(ecr-assets): expose property imageTag separately from imageUri in ECR assets  (#21582)
ECR assets has properties `imageUri` and `assetHash`. `assetHash` is used to generate the image tag, and `imageUri` is the repository and tag concatenated. However, parsing the tag from the URI is not possible since `imageUri` is a token. Additionally, `assetHash` and the image tag are not always identical since adding a synthesizer to the stack with a `dockerTagPrefix` would concatenate the prefix to the hash to make the tag.

It is now possible to get the tag by itself with `synthesizedTag`.

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*