fix(ec2): adds persist option to user data
To set the UserData for Windows instances to run on every instance
start, rather than only the first startup, requires adding
"<persist>true</persist>" to the end of the user data string.
Currently, users must use an escape hatch to hack the userdata of
windows instances to add this string. This PR makes it possible to
add this text directly on the UserData without using an escape hatch.