docs(cli): clarify app cli option (#21715)
The CLI documentation seen when running `--help` is misleading about the `--app` option being required. This option can be (and most often is) set by `cdk.json`, making the CLI option itself not strictly required when an app is needed. Additionally, no app value needs to be passed if the invoked command doesn't run the app (notably on `bootstrap` and `context`, possibly others). This PR clarifies these conditions

Fixes: #21564

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*