fix(core): feature flag values should be booleans
`FeatureFlags.isEnabled()` returns `boolean | undefined`, but if the
value of the feature flag is something other than a boolean, it will
actually return that type instead. For example,

```ts
node.setContext(someFeatureFlag, 'true');
const enabled = FeatureFlags.of(node).isEnabled(someFeatureFlag);
if (enabled) {
  // will work
}

if (enabled === true) {
  // will not work because enabled === 'true'
}
```

It looks like the only place this bug crops up is in the
`FunctionVersionUpgrade` aspect.

https://github.com/aws/aws-cdk/blob/8b0832334afee496355e7aeb684773d8f939f058/packages/@aws-cdk/aws-lambda/lib/function.ts#L1306