feat(cdk-cli-wrapper): add `progress` argument for cdk deploy
Adding the ability to specify `--progress` when using cdk `deploy`. I'm
setting the default to `events` as opposed to the default of `bar` in
the CLI. Since this library is currently only used in the integ-runner,
and if published will most likely be used in automation I think it makes
more sense to set progress to `events`.

This is really useful when running multiple integration tests in
parallel. If set to `bar` (CLI default) the logs will overlap/overwrite
each other which makes it difficult to troubleshoot.