feat(cdk-cli-wrapper): add `progress` argument for cdk deploy (#21762)
Adding the ability to specify `--progress` when using cdk `deploy`. I'm
setting the default to `events` as opposed to the default of `bar` in
the CLI. Since this library is currently only used in the integ-runner,
and if published will most likely be used in automation I think it makes
more sense to set progress to `events`.

This is really useful when running multiple integration tests in
parallel. If set to `bar` (CLI default) the logs will overlap/overwrite
each other which makes it difficult to troubleshoot.


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*