fix(integ): write assertion stack name to integ manifest
In #21646 we updated the `assertionStack` value to be the `node.id` of
the assertion stack since that is what the CDK CLI requires to deploy
the stack.

The change in #21646 fixed that bug, but it introduced a new one where the
assertion results are no longer read and reported on by the
integ-runner. The integ-runner reads the results of the assertions from
the stack outputs which are written to a file (with `cdk deploy --output
`assertion-results.json`). The outputs use the stack _name_ not the
_node.id_. As a result, the integ-runner was looking for outputs for an
invalid stack name.

This PR fixes that by:
- Adding `assertionStackName` property to the `integ.json` manifest
- Updates the integ-runner to use the `assertionsStackName` to lookup
  the assertion results.
- Add reporting on assertion results (previously only reported
  failures).
- Updates all integ snapshots which currently use assertions.