chore(docs): clarify use of isOrganizationTrail
If `isOrganizationTrail` is set to `true`, but the current AWS account
is _not_ the management account then CloudFormation will throw an error.

`You cannot configure or manage trails for an organization because this action requires being signed in with the management account.`

Also remove the integration test that was added for this feature since
in order to test this you would need access to an AWS organization
management account _and_ access to create/delete an organization trail.