fix(integ): write assertion stack name to integ manifest (#21809)
In #21646 we updated the `assertionStack` value to be the `node.id` of
the assertion stack since that is what the CDK CLI requires to deploy
the stack.

The change in #21646 fixed that bug, but it introduced a new one where the
assertion results are no longer read and reported on by the
integ-runner. The integ-runner reads the results of the assertions from
the stack outputs which are written to a file (with `cdk deploy --output
`assertion-results.json`). The outputs use the stack _name_ not the
_node.id_. As a result, the integ-runner was looking for outputs for an
invalid stack name.

This PR fixes that by:
- Adding `assertionStackName` property to the `integ.json` manifest
- Updates the integ-runner to use the `assertionsStackName` to lookup
  the assertion results.
- Add reporting on assertion results (previously only reported
  failures).
- Updates all integ snapshots which currently use assertions.


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*