chore(docs): clarify use of isOrganizationTrail (#21824)
If `isOrganizationTrail` is set to `true`, but the current AWS account
is _not_ the management account then CloudFormation will throw an error.

`You cannot configure or manage trails for an organization because this action requires being signed in with the management account.`

Also remove the integration test that was added for this feature since
in order to test this you would need access to an AWS organization
management account _and_ access to create/delete an organization trail.


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*