feat: compress aws-cdk-lib tablet file (#21854)
This is part of the Reduce Module Size [project](https://github.com/aws/aws-cdk/projects/15). Most of the heavy lifting is done in jsii, this PR simply turns on the feature.

I have tested this in our test pipeline and can confirm that the build output from there can be successfully ingested by contruct hub with transliterated examples in every language. Afaik, the tablet file included in the `aws-cdk-lib` package is _only_ for construct hub, because reproducing the tablet file would take ~3 hours.

Title is a `feat` so that it shows up in the changelog; there are no unit or integ tests that can be run to demonstrate it is working.

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*